+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.20.1] - 2025-12-11
+
+### Fixed
+
+- Plugin startup crash when PlayerCountPresence accessed services.discord before initialization
+- JDA "zip file closed" error during plugin shutdown by waiting for graceful thread termination
+
 ## [0.20.0] - 2025-12-11
 
 ### Added